requirements.txt
pandas==2.3.3
numpy==2.0.2
scipy==1.13.1
pyyaml==6.0.3
yfinance==1.2.0
streamlit==1.50.0
plotly==6.6.0
seaborn==0.13.2
matplotlib==3.9.4
pytest==8.4.2
